refactor: 패키지 분할 및 의존관계 수정
