+++ src/main/java/lotto/domain/prize/Prize.java
@@ -5,6 +5,7 @@ import lotto.domain.lottery.NumberChecker;
 import lotto.exception.LottoException;
 
 import static lotto.exception.ErrorMessage.BONUS_NUMBER_DUPLICATED;
+import static lotto.exception.ErrorMessage.NUMBER_OUT_OF_RANGE;
 
 public final class Prize extends NumberChecker {
     private final Lotto prizeNumbers;
@@ -16,7 +17,7 @@ public final class Prize extends NumberChecker {
             final int bonusNumber
     ) {
         this.prizeNumbers = prizeNumbers;
-        
+
         validateNumberRange(bonusNumber);
         validateDuplicatedBonusNumber(bonusNumber);
 
@@ -44,6 +45,12 @@ public final class Prize extends NumberChecker {
     private void validateDuplicatedBonusNumber(final int number) {
         if (prizeNumbers.hasNumber(number)) {
             throw LottoException.from(BONUS_NUMBER_DUPLICATED);
+        }
+    }
+
+    private void validateNumberRange(final Integer number) {
+        if (isOutOfRange(number)) {
+            throw LottoException.from(NUMBER_OUT_OF_RANGE);
         }
     }
 }
