+++ src/test/java/lotto/domain/lottery/ParserTest.java
@@ -1,5 +1,6 @@
 package lotto.domain.lottery;
 
+import lotto.service.Parser;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
